Move `Vertex` validation check to `HalfEdge`
This is preparation for some simplifications of `HalfEdge`/`Vertex` that
I'm currently working on.
